Module Gocodegen

type codegen_result = {

messages : Base.string;

channels : Base.string Base.Map.M(Names.ProtocolName).t;

invite_channels : Base.string Base.Map.M(Names.ProtocolName).t;

results : Base.string Base.Map.M(Names.ProtocolName).t;

impl : Base.string Base.Map.M(Names.LocalProtocolName).t;

callbacks : Base.string Base.Map.M(Names.LocalProtocolName).t;

protocol_setup : Base.string Base.Map.M(Names.ProtocolName).t;

entry_point : Base.string;

}

Record containing the implementation of all the components of the Go protocol implementation

val show_codegen_result : codegen_result -> Names.ProtocolName.t -> Names.RootDirName.t -> Base.string

Convert the code generation result to a string, showing the contents of each generated implementation file

val gen_code : Names.RootDirName.t -> Names.ProtocolName.t -> Mpst.Gtype.nested_t -> Mpst.Ltype.nested_t -> codegen_result

Function which generates the Go implementation for the entry-point protocol and all the protocols in the Scribble module

val ensure_unique_identifiers : Mpst.Gtype.nested_t -> Base.unit

Function which verifies that the Scribble module satisfies the constraints needed in order to be able to generate implementation. Some of these constraints may be too restrictive right now, but they should be easy to relax

val generate_go_code : Syntaxtree.Syntax.scr_module -> protocol:Names.ProtocolName.t -> out_dir:Base.string -> go_path:Base.string Base.option -> Base.string

generate_code module protocol out_dir go_path generates Golang implementation for protocol. The protocol implementation designed to be a subpackage within a project. out_dir is the path from the root of the project until the package inside which the protocol implementation (subpackage) should be generated - it is needed to generate imports. go_path is the path to the project root, which can optionally be provided in order to write the implementation to the file system.