Module Goenvs

module ImportsEnv : sig ... end

Environment which keeps track of the package imports which which need to be generated. Importing new packages multiple times doesn't change the environment, and importing two packages with the same name will result in different package aliases being generated

module MessagesEnv : sig ... end

Environment for generating structs for the different messages exchanged during a protocol. Every unique message label will correspond to a different struct

module EnumNamesEnv : sig ... end

Environment used to keep track of all the enums which have been created within the 'callbacks' package in order to ensure that the generated enum names are unique

module ChannelEnv : sig ... end

Environment to generate the channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

module InviteEnv : sig ... end

Environment to generate the invitation channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

module ProtocolSetupEnv : sig ... end

Environment used to keep track of the structs containing the channels which will be used during the setup of each protocol

module ProtocolSetupGen : sig ... end

Environment for creating all of the channel variables and structs during the setup of a protocol

module CallbacksEnv : sig ... end

Environment used to generate the callbacks interface for a role. It keeps track of all the callbacks that a role's implemenatation needs and generates the callbacks source file for that role, ensuring that the names of the generated callbacks are unique

type protocol_env = {

channel_imports : ImportsEnv.t;

invite_imports : ImportsEnv.t;

callback_enum_names : EnumNamesEnv.t;

}

module LTypeCodeGenEnv : sig ... end

Wrapper around all the different environments used during the code generation