Module Goenvs.CallbacksEnv

Environment used to generate the callbacks interface for a role. It keeps track of all the callbacks that a role's implemenatation needs and generates the callbacks source file for that role, ensuring that the names of the generated callbacks are unique

type t

val new_send_callback : t -> Names.LabelName.t -> Mpst.Gtype.payload Base.list -> Names.RoleName.t -> t * Names.CallbackName.t

Add callback for returning the payloads for the message which will be sent in a labelled message exchange

val new_recv_callback : t -> Names.LabelName.t -> Mpst.Gtype.payload Base.list -> Names.RoleName.t -> t * Names.CallbackName.t

Add callback which takes in the message which was received in labelled message exchange

val new_protocol_setup_callback : t -> Names.ProtocolName.t -> t * Names.CallbackName.t

Add callback for setting up a protocol call

val new_convert_env_callback : t -> Names.LocalProtocolName.t -> t * Names.CallbackName.t

Add callback for creating the environment for the new role that the participant will carry out in a protocol call

val new_choice_callback : t -> Names.RoleName.t -> Names.LabelName.t Base.list -> Names.LocalProtocolName.t -> t * Names.CallbackName.t * Names.EnumName.t Base.list

Add callback for returning the choice that the role will make

val new_protocol_result_callback : t -> Names.LocalProtocolName.t -> Names.ProtocolName.t -> Names.RoleName.t -> t * Names.CallbackName.t

Add callback for receiving the result which was returned after finishing a protocol call

val add_done_callback : t -> Names.ProtocolName.t -> Names.RoleName.t -> Base.bool -> t * Names.CallbackName.t

Add callback for finishing the protocol's execution: Done(). This callback is only added once, and depending on whether the role is a dynamic participant or not, the function will be void or return a result

val get_enum_names_env : t -> EnumNamesEnv.t

Return the enum names environment which stores all the choice enums generated in the callbacks package so far

val gen_callbacks_file : t -> Names.LocalProtocolName.t -> Base.bool -> Base.string

Generate callbacks source file for the role

val create : Names.RootDirName.t -> EnumNamesEnv.t -> t