Module Goenvs.ChannelEnv

Environment to generate the channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

type t

val get_or_add_channel : t -> (Names.RoleName.t * Names.PayloadTypeName.t Base.option * Base.bool) -> t * (Names.ChannelName.t * Base.string)

Generate new struct field for a channel to send or receive a payload or message label

val gen_channel_struct : t -> Base.string

Generate role channel struct declaration

val get_channel_imports : t -> ImportsEnv.t

Return the imports environment with the imports needed for the channels file which have been aggregated

val struct_name : t -> Names.ChannelStructName.t

Returns the name of the channel struct

val create : Names.RoleName.t -> Names.ProtocolName.t -> Names.ProtocolName.t -> ImportsEnv.t -> t