Module Goenvs.EnumNamesEnv

Environment used to keep track of all the enums which have been created within the 'callbacks' package in order to ensure that the generated enum names are unique

type t

val create : Base.unit -> t

val new_enum : t -> Names.LocalProtocolName.t -> Names.LabelName.t Base.list -> t * Names.EnumTypeName.t * Names.EnumName.t Base.list

Generate unique names for the enum type and enum values based on the local protocol and the labels of the first interactions in the choice