Module Goenvs.ImportsEnv

Environment which keeps track of the package imports which which need to be generated. Importing new packages multiple times doesn't change the environment, and importing two packages with the same name will result in different package aliases being generated

type t

val import_messages : t -> t * Names.PackageName.t

Import package 'messages/<protocol>'

val import_channels : t -> Names.ProtocolName.t -> t * Names.PackageName.t

Import package 'channels/<protocol>'

val import_results : t -> Names.ProtocolName.t -> t * Names.PackageName.t

Import package 'results/<protocol>'

val import_invitations : t -> t * Names.PackageName.t

Import package 'invitations'

val import_callbacks : t -> t * Names.PackageName.t

Import package 'callbacks'

val import_roles : t -> t * Names.PackageName.t

Import package 'roles'

val import_sync : t -> t * Names.PackageName.t

Import package 'sync'

val generate_imports : t -> Base.string

Generate all the Go import statements from the given environment

val create : Names.RootDirName.t -> t

Create a new empty environment. Root directory is used to generate relative imports