Module Goenvs.InviteEnv

Environment to generate the invitation channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

type t

type role_channel_field = Names.InviteChannelName.t * (Names.PackageName.t * Names.ChannelStructName.t)

type role_invite_channel_field = Names.InviteChannelName.t * Names.InviteChannelStructName.t

val get_or_add_send_invitation_channels : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.RoleName.t -> Names.ProtocolName.t -> t * role_channel_field * role_invite_channel_field

Generate, if they don't exist already, two new struct field for sending the invitations to a role (potentially itself). Invitations consist of the channel struct and invitation struct that the role will need to communicate in the nested protocol

val get_or_add_recv_invitation_channels : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.RoleName.t -> Names.ProtocolName.t -> t * role_channel_field * role_invite_channel_field

Generate, if they don't exist already, two new struct field for receiving the invitations from the role which is initiating a protocol call - this should not be called if the role itself is initiating the protocol call. Invitations consist of the channel struct and invitation struct that the role will need to communicate in the nested protocol

val gen_invite_channel_struct : t -> Base.string

Generate invitation channel struct

val get_invite_imports : t -> ImportsEnv.t

Return imports environment with all the imports for the invitations source file aggregated so far

val struct_name : t -> Names.InviteChannelStructName.t

Return name of the invitation channel struct

val create : Names.LocalProtocolName.t -> ImportsEnv.t -> t