Module Goenvs.MessagesEnv

Environment for generating structs for the different messages exchanged during a protocol. Every unique message label will correspond to a different struct

type t

val create : Names.ProtocolName.t -> t

val add_message_enum : t -> Names.LabelName.t -> t

Add enum for message label if it doesn't already exist.

val add_invitation_enum : t -> Names.ProtocolName.t -> Names.RoleName.t Base.list -> t

Add enum for protocol call if it doesn't already exist

val get_message_enum : t -> Names.LabelName.t -> Names.EnumName.t

Get enum value for message label

val get_enum_type_name : t -> Names.EnumTypeName.t

Get name of enum type

val get_invitation_enum : t -> Names.ProtocolName.t -> Names.RoleName.t Base.list -> Names.EnumName.t

Get enum value for protocol call

val generate_messages_file : t -> Base.string

Generate Go source file containing the message structs defined in the environment