Module Goenvs.ProtocolSetupEnv

Environment used to keep track of the structs containing the channels which will be used during the setup of each protocol

type t

val get_setup_channel_struct : t -> Names.ProtocolName.t -> Names.InviteChannelStructName.t * Names.InviteChannelName.t Base.list

Return the struct name and fields of the channel setup struct for the given protocol

val get_setup_invite_struct : t -> Names.ProtocolName.t -> Names.InviteChannelStructName.t * Names.InviteChannelName.t Base.list

Return the struct name and fields of the invitation setup struct for the given protocol

val get_setup_function_name : t -> Names.ProtocolName.t -> Names.FunctionName.t

Generate name for a protocol's setup function

val gen_setup_channel_struct : t -> Names.PackageName.t -> Names.ProtocolName.t -> Base.string

Generate setup channel struct declaration for the given protocol

val gen_setup_invite_struct : t -> Names.ProtocolName.t -> Base.string

Generate setup invitation struct declaration for the given protocol

val create : Mpst.Ltype.local_proto_name_lookup -> Mpst.Ltype.nested_t -> Mpst.Gtype.nested_t -> t

Create new environment with all the information about the structs