Module Mpst__Gtype

type payload =

| PValue of Names.VariableName.t Base.option * Mpst.Expr.payload_type

| PDelegate of Names.ProtocolName.t * Names.RoleName.t

val equal_payload : payload -> payload -> Ppx_deriving_runtime.bool

val pp_payload : Ppx_deriving_runtime.Format.formatter -> payload -> Ppx_deriving_runtime.unit

val show_payload : payload -> Ppx_deriving_runtime.string

val compare_payload : payload -> payload -> Ppx_deriving_runtime.int

val sexp_of_payload : payload -> Ppx_sexp_conv_lib.Sexp.t

type message = {

label : Names.LabelName.t;

payload : payload Base.list;

}

val equal_message : message -> message -> Ppx_deriving_runtime.bool

val pp_message : Ppx_deriving_runtime.Format.formatter -> message -> Ppx_deriving_runtime.unit

val show_message : message -> Ppx_deriving_runtime.string

val compare_message : message -> message -> Ppx_deriving_runtime.int

val sexp_of_message : message -> Ppx_sexp_conv_lib.Sexp.t

val equal_pvalue_payload : payload -> payload -> Base.bool

val typename_of_payload : payload -> Names.PayloadTypeName.t

type rec_var = {

rv_name : Names.VariableName.t;
Variable Name

rv_roles : Names.RoleName.t Base.list;
Which roles know this variable

rv_ty : Mpst.Expr.payload_type;
What type does the variable carry

rv_init_expr : Mpst.Expr.t;
What is the initial expression assigned at the beginning of recursion

}

Recursion variable

val equal_rec_var : rec_var -> rec_var -> Ppx_deriving_runtime.bool

val sexp_of_rec_var : rec_var -> Ppx_sexp_conv_lib.Sexp.t

type t =

| MessageG of message * Names.RoleName.t * Names.RoleName.t * t
MessageG (msg, sender, receiver, t) starts by sending message msg from sender to receiver and continues as t

| MuG of Names.TypeVariableName.t * rec_var Base.list * t

| TVarG of Names.TypeVariableName.t * Mpst.Expr.t Base.list * t Base.Lazy.t

| ChoiceG of Names.RoleName.t * t Base.list
ChoiceG (name, ts) expresses a choice located at participant name between the ts

| EndG
Empty global type

| CallG of Names.RoleName.t * Names.ProtocolName.t * Names.RoleName.t Base.list * t
CallG (caller, protocol, participants, t) - caller calls protocol, inviting participants to carry out the roles in protocol (dynamic roles in nested protocols are not included)

The type of global types

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type nested_global_info = {

static_roles : Names.RoleName.t Base.list;

dynamic_roles : Names.RoleName.t Base.list;

nested_protocol_names : Names.ProtocolName.t Base.list;

gtype : t;

}

type nested_t = nested_global_info Base.Map.M(Names.ProtocolName).t

val show : t -> Base.string

Provides a textual representation of a global type

val show_nested_t : nested_t -> Base.string

Provides a textual representation of a global type with nested protocols

val call_label : Names.RoleName.t -> Names.ProtocolName.t -> Names.RoleName.t Base.list -> Names.LabelName.t

Generates a unique label for a protocol call based on the caller, the protocol called and the participants involved

val of_protocol : Syntaxtree.Syntax.global_protocol -> t

Turn a raw protocol (from the parser) into a global type, optional argument refined determines whether refinement types are enabled.

val nested_t_of_module : Syntaxtree.Syntax.scr_module -> nested_t

Turn scribble module (from the parser) into a nested global type

val normalise : t -> t

Normalise a global type. This mainly collapses nested choice on the same participant and unfolds fixpoints

val normalise_nested_t : nested_t -> nested_t

Apply normalisation to all protocols in nested_t

val validate_refinements_exn : t -> Base.unit

Validate refinements in the given global type, requires RefinementTypes pragma

val show_rec_var : rec_var -> Base.string

Convert rec_var to string