Module Mpst__Ltype

Local type management

type t =

| RecvL of Mpst.Gtype.message * Names.RoleName.t * t
RecvL (msg, name, t) waits for message msg from name and continues as t

| SendL of Mpst.Gtype.message * Names.RoleName.t * t
SendL (msg, name, t) sends message msg to name and continues as t

| ChoiceL of Names.RoleName.t * t Base.list
ChoiceL (name, ts) is a choice (internal or external) from name between the ts

| TVarL of Names.TypeVariableName.t * Mpst.Expr.t Base.list
Recursive variable

| MuL of Names.TypeVariableName.t * (Base.bool * Mpst.Gtype.rec_var) Base.list * t
Fixpoint

| EndL
Empty type

| InviteCreateL of Names.RoleName.t Base.list * Names.RoleName.t Base.list * Names.ProtocolName.t * t
Send invitations to existing roles and set up/create dynamic pariticipants

| AcceptL of Names.RoleName.t * Names.ProtocolName.t * Names.RoleName.t Base.list * Names.RoleName.t Base.list * Names.RoleName.t * t
accept role'@Proto(roles...; new roles'...) from X; t

| SilentL of Names.VariableName.t * Mpst.Expr.payload_type * t
Used with refinement types to indicate knowledge obtained via a global protocol.

Local types.

module LocalProtocolId : sig ... end

Unique id identifying a local protocol

type nested_t = (Names.RoleName.t Base.list * t) Base.Map.M(LocalProtocolId).t

Mapping of local protocol id to the protocol's roles and local type

val show : t -> Base.string

Converts a local type to a string.

val show_nested_t : nested_t -> Base.string

val project : Names.RoleName.t -> Mpst.Gtype.t -> t

Project a global type into a particular role.

val project_nested_t : Mpst.Gtype.nested_t -> nested_t

Generate the local protocols for a given global_t

val ensure_unique_tvars : nested_t -> nested_t

Ensure that all the local variables in each local protocol are globally unique, renaming variables to resolve conflicts

type local_proto_name_lookup = Names.LocalProtocolName.t Base.Map.M(LocalProtocolId).t

Mapping from local protocol ids to their unique local protocol names

val build_local_proto_name_lookup : nested_t -> local_proto_name_lookup

Builds a map containing the unique string representations for the unique local protocol ids

val show_lookup_table : local_proto_name_lookup -> Base.string

Converts a local protocol name lookup table to a string

val lookup_local_protocol : local_proto_name_lookup -> Names.ProtocolName.t -> Names.RoleName.t -> Names.LocalProtocolName.t

Return the unique local protocol name for a (role, protocol) pair

val lookup_protocol_id : local_proto_name_lookup -> LocalProtocolId.t -> Names.LocalProtocolName.t

Look up the unique name for a local protocol id