Module Nuscrlib.Gocodegen

type codegen_result = {

messages : (Names.ProtocolName.t, Base.string, Nuscrlib.Names.ProtocolName.comparator_witness) Base.Map.t;

channels : (Names.ProtocolName.t, Base.string, Nuscrlib.Names.ProtocolName.comparator_witness) Base.Map.t;

invite_channels : (Names.ProtocolName.t, Base.string, Nuscrlib.Names.ProtocolName.comparator_witness) Base.Map.t;

results : (Names.ProtocolName.t, Base.string, Nuscrlib.Names.ProtocolName.comparator_witness) Base.Map.t;

impl : (Names.LocalProtocolName.t, Base.string, Nuscrlib.Names.LocalProtocolName.comparator_witness) Base.Map.t;

callbacks : (Names.LocalProtocolName.t, Base.string, Nuscrlib.Names.LocalProtocolName.comparator_witness) Base.Map.t;

protocol_setup : (Names.ProtocolName.t, Base.string, Nuscrlib.Names.ProtocolName.comparator_witness) Base.Map.t;

entry_point : Base.string;

}

Record containing the implementation of all the components of the Go protocol implementation

val show_codegen_result : codegen_result -> Names.ProtocolName.t -> Names.RootDirName.t -> Base.string

Convert the code generation result to a string, showing the contents of each generated implementation file

val gen_code : Names.RootDirName.t -> Names.ProtocolName.t -> Gtype.global_t -> Ltype.local_t -> codegen_result

Function which generates the Go implementation for the entry-point protocol and all the protocols in the Scribble module

val ensure_unique_identifiers : Gtype.global_t -> Base.unit

Function which verifies that the Scribble module satisfies the constraints needed in order to be able to generate implementation. Some of these constraints may be too restrictive right now, but they should be easy to relax