Module Goenvs.InviteEnv

Environment to generate the invitation channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

type t

val new_send_role_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.RoleName.t -> Names.ProtocolName.t -> t * Names.InviteChannelName.t * Names.ChannelStructName.t

Generate new struct field for sending a role channel struct (first part of the invitation)

val new_send_invite_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> t * Names.InviteChannelName.t * Names.InviteChannelStructName.t

Generate new struct field for sending a invitation channel struct (second part of the invitation)

val send_self_role_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.InviteChannelName.t * Names.ChannelStructName.t

Return struct field name of the channel used to send the role channel struct invitation to itself. When a role receives an invitation, if the invitation was sent by itself then no new channel fields should be created. The same channel used to send the invitation should be returned

val send_self_invite_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.InviteChannelName.t * Names.InviteChannelStructName.t

Return struct field name of the channel used to send the invitation channel struct invitation to itself. When a role receives an invitation, if the invitation was sent by itself then no new channel fields should be created. The same channel used to send the invitation should be returned

val new_recv_role_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> Names.RoleName.t -> Names.ProtocolName.t -> t * Names.InviteChannelName.t * Names.ChannelStructName.t

Generate new channel field to accept the role channel struct invitation

val new_recv_invite_channel : t -> Names.RoleName.t -> Names.LocalProtocolName.t -> t * Names.InviteChannelName.t * Names.InviteChannelStructName.t

Generate new channel field to accept the invitation channel struct invitation

val gen_invite_channel_struct : t -> Base.string

Generate invitation channel struct

val get_invite_imports : t -> ImportsEnv.t

Return imports environment with all the imports for the invitations source file aggregated so far

val struct_name : t -> Names.InviteChannelStructName.t

Return name of the invitation channel struct

val create : Names.LocalProtocolName.t -> ImportsEnv.t -> t