Module Nuscrlib.Gtype

type payload =

| PValue of Names.VariableName.t Base.option * Names.PayloadTypeName.t

| PDelegate of Names.ProtocolName.t * Names.RoleName.t

val equal_payload : payload -> payload -> Ppx_deriving_runtime.bool

val pp_payload : Ppx_deriving_runtime.Format.formatter -> payload -> Ppx_deriving_runtime.unit

val show_payload : payload -> Ppx_deriving_runtime.string

val compare_payload : payload -> payload -> Ppx_deriving_runtime.int

val sexp_of_payload : payload -> Ppx_sexp_conv_lib.Sexp.t

type message = {

label : Names.LabelName.t;

payload : payload Base.list;

}

val equal_message : message -> message -> Ppx_deriving_runtime.bool

val pp_message : Ppx_deriving_runtime.Format.formatter -> message -> Ppx_deriving_runtime.unit

val show_message : message -> Ppx_deriving_runtime.string

val compare_message : message -> message -> Ppx_deriving_runtime.int

val sexp_of_message : message -> Ppx_sexp_conv_lib.Sexp.t

val equal_pvalue_payload : payload -> payload -> Base.bool

type t =

| MessageG of message * Names.RoleName.t * Names.RoleName.t * t
MessageG (msg, sender, receiver, t) starts by sending message msg from sender to receiver and continues as t

| MuG of Names.TypeVariableName.t * t
Fixpoint

| TVarG of Names.TypeVariableName.t * t Base.Lazy.t
Recursive variable

| ChoiceG of Names.RoleName.t * t Base.list
ChoiceG (name, ts) expresses a choice located at participant name between the ts

| EndG
Empty global type

| CallG of Names.RoleName.t * Names.ProtocolName.t * Names.RoleName.t Base.list * t
CallG (caller, protocol, participants, t) - caller calls protocol, inviting participants to carry out the roles in protocol (dynamic roles in nested protocols are not included)

The type of global types

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type global_t = ((Names.RoleName.t Base.list * Names.RoleName.t Base.list) * Names.ProtocolName.t Base.list * t) Base.Map.M(Names.ProtocolName).t

Mapping of protocol name to the roles ('static' participants, dynamic participants) participating in the protocol, the names of the nested protocols defined inside it and its global type

val show : t -> Base.string

Provides a textual representation of a global type

val show_global_t : global_t -> Base.string

Provides a textual representation of a global type with nested protocols

val call_label : Names.RoleName.t -> Names.ProtocolName.t -> Names.RoleName.t Base.list -> Names.LabelName.t

Generates a unique label for a protocol call based on the caller, the protocol called and the participants involved

val of_protocol : Syntax.global_protocol -> t

Turn a raw protocol (from the parser) into a global type

val global_t_of_module : Syntax.scr_module -> global_t

Turn scribble module (from the parser) into a global type

val global_t_of_ast : Syntax.scr_module -> global_t

Turn scribble ast (from the parser) into a global type

val normalise : t -> t

Normalise a global type. This mainly collapses nested choice on the same participant and unfolds fixpoints

val normalise_global_t : global_t -> global_t

Apply normalisation to all protocols in global_t

val replace_recursion_with_nested_protocols : global_t -> global_t

Replace the MuG type with explicit calls to nested protocols