Module Nuscrlib.Ltype

type t =

| RecvL of Syntax.message * Syntax.name * t
RecvL (msg, name, t) waits for message msg from name and continues as t

| SendL of Syntax.message * Syntax.name * t
SendL (msg, name, t) sends message msg to name and continues as t

| ChoiceL of Syntax.name * t list
ChoiceL (name, ts) is a choice (internal or external) from name between the ts

| TVarL of Syntax.name
Recursive variable

| MuL of Syntax.name * t
Fixpoint

| EndL
Empty type

Local types.

val show : t -> string

Converts a local type to a string.

val project : Syntax.name -> Gtype.t -> t

Project a global type into a particular role.