Module type Names.TaggedName

A TaggedName is a module type to represent names of different kinds. One can use Make () to create a new namespace module, so that different namespaces can be properly distinguished at the type level

include UntaggedName

type t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_string : Base.string -> t

Creates a name from a string, with a default location (unspecified)

val rename : t -> Base.string -> t

Change the identifier, without changing the location

val update : t -> f:( Base.string -> Base.string ) -> t

Update the identifier with function ~f, without changing the location

val user : t -> Base.string

Prints the identifier in a user-friendly format, i.e. without the location

val where : t -> Loc.t

Gives the location of the identifier

val create : Base.string -> Loc.t -> t

Creates a name from an identifier and a location

val hash : t -> Base.int

Hash (identifier only, location ignored)

include Base.Comparable.S with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base__.Or_error.t

type comparator_witness

val comparator : ( t, comparator_witness ) Base__.Comparator.comparator

val of_other_name : (module UntaggedName with type t = 'a) -> 'a -> t

Allows conversion between different tags