Module Nuscrlib__Goenvs.ChannelEnv

Environment to generate the channel struct for a role in a protocol. The environment keeps track of all the field names and types in order to generate the struct declaration

type t

val new_channel : t -> Nuscrlib.Names.RoleName.t -> Nuscrlib.Names.LabelName.t -> t * Nuscrlib.Names.ChannelName.t * Nuscrlib.Names.MessageStructName.t

Generate new struct field given the label of the message exchanged

val gen_channel_struct : t -> Base.string

Generate role channel struct declaration

val get_channel_imports : t -> ImportsEnv.t

Return the imports environment with the imports needed for the channels file which have been aggregated

val struct_name : t -> Nuscrlib.Names.ChannelStructName.t

Returns the name of the channel struct

val create : Nuscrlib.Names.RoleName.t -> Nuscrlib.Names.ProtocolName.t -> ImportsEnv.t -> t