Module Nuscrlib__Goenvs.MessagesEnv

Environment for generating structs for the different messages exchanged during a protocol. Every unique message label will correspond to a different struct

type t

val create : Base.unit -> t

val add_message_struct : t -> Nuscrlib.Names.LabelName.t -> Nuscrlib.Gtype.payload Base.list -> t * Nuscrlib.Names.MessageStructName.t

Add struct for labeled message if it doesn't already exist. Assumes message labels are used consistently throughout the protocol

val generate_messages_file : t -> Nuscrlib.Names.PackageName.t -> Base.string

Generate Go source file containing the message structs defined in the environment