Module Nuscrlib__Goenvs.ProtocolSetupGen

Environment for creating all of the channel variables and structs during the setup of a protocol

type t

val create : Nuscrlib.Names.RootDirName.t -> Nuscrlib.Names.RoleName.t Base.list -> Nuscrlib.Names.ProtocolName.t -> Nuscrlib.Ltype.local_proto_name_lookup -> t

val generate_setup_channels : Nuscrlib.Gtype.global_t -> Nuscrlib.Ltype.local_proto_name_lookup -> t -> ImportsEnv.t * Nuscrlib.Names.VariableName.t Base.Map.M(Nuscrlib.Names.RoleName).t * Nuscrlib.Names.VariableName.t Base.Map.M(Nuscrlib.Names.RoleName).t * Base.string

Generate the assignments for all of the channels needed in a protocol and create all the channel and invitation structs for each role with them. Return the names of the variables which contain the channel and invitation structs for each role as well as the setup implementation.