Module Nuscrlib__Gtype

type t =

| MessageG of Nuscrlib.Syntax.message * Nuscrlib.Syntax.name * Nuscrlib.Syntax.name * t
MessageG (msg, sender, receiver, t) starts by sending message msg from sender to receiver and continues as t

| MuG of Nuscrlib.Syntax.name * t
Fixpoint

| TVarG of Nuscrlib.Syntax.name
Recursive variable

| ChoiceG of Nuscrlib.Syntax.name * t list
ChoiceG (name, ts) expresses a choice located at participant name between the ts

| EndG
Empty global type

The type of global types

val show : t -> string

Provides a textual representation of a global type

val of_protocol : Nuscrlib.Syntax.raw_global_protocol Nuscrlib.Syntax.located -> t

Turn a raw protocol (from the parser) into a global type

val normalise : t -> t

Normalise a global type. This mainly collapses nested choice on the same participant and unfolds fixpoints