Module Nuscrlib__Ltype

Local type management

type t =

| RecvL of Nuscrlib.Syntax.message * Nuscrlib.Syntax.name * t
RecvL (msg, name, t) waits for message msg from name and continues as t

| SendL of Nuscrlib.Syntax.message * Nuscrlib.Syntax.name * t
SendL (msg, name, t) sends message msg to name and continues as t

| ChoiceL of Nuscrlib.Syntax.name * t list
ChoiceL (name, ts) is a choice (internal or external) from name between the ts

| TVarL of Nuscrlib.Syntax.name
Recursive variable

| MuL of Nuscrlib.Syntax.name * t
Fixpoint

| EndL
Empty type

Local types.

val show : t -> string

Converts a local type to a string.

val project : Nuscrlib.Gtype.t -> Nuscrlib.Syntax.name list -> Nuscrlib.Syntax.name -> t

Project a global type into a particular role.