Module Syntaxtree.Symtable

type protocol_decl = {

proto_name : Base.string;

all_roles : Base.string Base.list;

split_decl : Base.string Base.list * Base.string Base.list;

loc : Loc.t;

}

Record containing a protocol's signature

val show_roles : (Base.String.t list * Base.String.t list) -> Base.String.t

val show_protocol_decl : protocol_decl -> Base.String.t

type symbol_table = {

protocol : Base.string;

table : protocol_decl Base.Map.M(Base.String).t;

parent : symbol_table Base.option;

}

Symbol table type

val pp_symbol_table : Ppx_deriving_runtime.Format.formatter -> symbol_table -> Ppx_deriving_runtime.unit

val show_symbol_table : symbol_table -> Ppx_deriving_runtime.string

val name_with_prefix : Base.String.t -> Base.String.t -> Base.String.t

Generate nested protocol names based on the name of the parent protocol

val decl_from_protocol : Base.String.t -> Syntax.global_protocol -> protocol_decl

Generate a protocol's declaration

val build_symbol_table : Base.string -> Syntax.global_protocol Base.list -> symbol_table Base.option -> symbol_table

Build symbol table for a given protocol

val lookup_protocol : symbol_table -> Syntax.N.t -> protocol_decl

Recursively look up a protocol in the symbol table hierarchy