Module Nuscrlib.Gtype

type t =

| MessageG of Syntax.message * Syntax.name * Syntax.name * t
MessageG (msg, sender, receiver, t) starts by sending message msg from sender to receiver and continues as t

| MuG of Syntax.name * t
Fixpoint

| TVarG of Syntax.name
Recursive variable

| ChoiceG of Syntax.name * t list
ChoiceG (name, ts) expresses a choice located at participant name between the ts

| EndG
Empty global type

The type of global types

val show : t -> string

Provides a textual representation of a global type

val of_protocol : Syntax.raw_global_protocol Loc.located -> t

Turn a raw protocol (from the parser) into a global type

val normalise : t -> t

Normalise a global type. This mainly collapses nested choice on the same participant and unfolds fixpoints